+++ FraudProtection/src/main/java/com/paypal/android/fraudprotection/PayPalDataCollector.kt
@@ -32,7 +32,7 @@ class PayPalDataCollector internal constructor(
     constructor(config: CoreConfig) : this(config, MagnesSDK.getInstance(), UUIDHelper())
 
     /**
-     * Gets a Client Metadata ID at the time of payment activity. Once a user initiates a payment
+     * Collects device data at the time of payment activity. Once a user initiates a payment
      * from their device, PayPal uses the Client Metadata ID to verify that the payment is
      * originating from a valid, user-consented device and application. This helps reduce fraud and
      * decrease declines. This method MUST be called prior to initiating a pre-consented payment (a
@@ -46,7 +46,7 @@ class PayPalDataCollector internal constructor(
      * @return clientMetadataId Your server will send this to PayPal
      */
     @JvmOverloads
-    fun getClientMetadataId(
+    fun collectDeviceData(
         context: Context,
         clientMetadataId: String? = null,
         additionalData: HashMap<String, String>? = null